+++ packages/antd/modules/widgets/value/Slider.jsx
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import PropTypes from "prop-types";
 import { Slider, InputNumber, Col } from "antd";
 import { Utils } from "@react-awesome-query-builder/ui";
-const { useOnPropsChanged } = Utils.ReactUtils;
+const { useOnPropsChanged, pureShouldComponentUpdate } = Utils.ReactUtils;
 const __isInternal = true; //true to optimize render
 
 export default class SliderWidget extends Component {
@@ -34,6 +34,7 @@ export default class SliderWidget extends Component {
 
   constructor(props) {
     super(props);
+    this.pureShouldComponentUpdate = pureShouldComponentUpdate(this);
     useOnPropsChanged(this);
 
     this.state.internalValue = props.value;
@@ -53,12 +54,16 @@ export default class SliderWidget extends Component {
 
   tipFormatter = (val) => (val != undefined ? val.toString() : undefined);
 
-  UNSAFE_componentWillUpdate(nextProps, nextState) {
-    // RHL fix
-    if (this.props.cacheBusterProp && __isInternal) {
-      nextState.internalValue = this.state.internalValue;
+  shouldComponentUpdate = (nextProps, nextState) => {
+    const should = this.pureShouldComponentUpdate(nextProps, nextState);
+    if (should) {
+      // RHL fix
+      if (this.props.cacheBusterProp && __isInternal) {
+        nextState.internalValue = this.state.internalValue;
+      }
     }
-  }
+    return should;
+  };
 
   render() {
     const {config, placeholder, customProps, value,  min, max, step, marks, readonly, valueError} = this.props;