+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - Fix isNull roundtrip (PR #887) (issue #886)
   - Fix `BootstrapFieldSelect` for fields with 2+ level nesting (PR #898) (issue #868)
   - Fix `UNSAFE_componentWillReceiveProps` (PR #901) (issue #390)
+  - Fix bugs with `React.StrictMode` (PR #902) (issue #789)
 - 6.1.2
   - Fix typings for `children1`: tuple -> array (PR #885) (issue #881)
   - Fix compare fields for case when type != widget.type (PR #875) (issue #758)